Don't delete reply drafts when deleting a body top/bottom field.
Due to some bad logic, creating an empty comment saving it for the body
top/bottom replies to a review would trigger the full deletion of a
draft, possibly losing a lot of work. This happened because the logic
attempts to delete a comment if its text was set to empty, but when
replying to a review directly, it would delete the review instead.

Now the logic only attempts to do a straight deletion if it knows it's
not replying to the body top/bottom fields. In that case, it simply
calls `_resetState()`, which will reset the state for the particular
field and then triggers the checks to see if the review is now empty.

Testing Done:
Created and saved a reply to a diff comment, and then created a reply to
a body top field, leaving it blank and hitting OK. Saw that only that
particular reply disappeared, rather than the draft being discarded as
before.

Tested the other direction and verified that the behavior was as expected.

Cleared out both replies (tried first with the body_top being the last
one cleared out, and second with the comment being cleared out last) and
verified the draft disappeared when there was nothing remaining.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/10039/

+++ reviewboard/static/rb/js/reviewRequestPage/models/reviewReplyEditorModel.js
@@ -116,18 +116,28 @@ RB.ReviewRequestPage.ReviewReplyEditor = Backbone.Model.extend({
      */
     resetStateIfEmpty: function() {
         var text = this.get('text'),
-            replyObject = this.get('replyObject');
+            replyObject,
+            contextType;
 
         if (text.strip() !== '') {
             return;
         }
 
+        replyObject = this.get('replyObject');
+
         if (!replyObject || replyObject.isNew()) {
             this._resetState();
         } else {
-            replyObject.destroy({
-                success: this._resetState
-            }, this);
+            contextType = this.get('contextType');
+
+            if (contextType === 'body_top' ||
+                contextType === 'body_bottom') {
+                this._resetState(true);
+            } else {
+                replyObject.destroy({
+                    success: this._resetState
+                }, this);
+            }
         }
     },
